+++ src/operations/booleans/subtract.js
@@ -1,5 +1,5 @@
-const flatten = require('../utils/flatten')
-const areAllShapesTheSameType = require('../utils/areAllShapesTheSameType')
+const flatten = require('../../utils/flatten')
+const areAllShapesTheSameType = require('../../utils/areAllShapesTheSameType')
 
 const {geom2, geom3, path2} = require('../../geometry')
