+++ src/operations/booleans/to3DWalls.js
@@ -1,4 +1,4 @@
-const vec3 = require('../../math/vec3')
+const {vec3} = require('../../math')
 
 const {geom2, geom3, poly3} = require('../../geometry')
